Return 404 when document requested doesn't belong to patient
Details: To prevent someone from discovering that a patient has documents in the system by querying random document IDs with the wrong MRN and looking at the error message returned when they try to access someone else's document, change the HTTP status returned from 403 - Forbidden to 404 - Not Found, which is the same code returned when trying to query a document ID which does not exist. However, the server-side log will still record if someone tries to request a document which doesn't belong to him/her.

#13764 - As a developer, I will have implemented the Upload Clinical Document UI and integrated with backend